完成查询功能和编辑功能

+++ router/rule/model.go
@@ -4,6 +4,7 @@ import (
 	"database/sql"
 	"go-short-url/util"
 	"log"
+	"strings"
 	"time"
 )
 
@@ -23,7 +24,7 @@ type Rule struct {
 	UpdateTime time.Time  `json:"updateTime"`
 }
 
-func (rule Rule) Insert(db *sql.DB) (r *Rule, err error) {
+func (rule *Rule) Insert(db *sql.DB) error {
 	stmt, err := db.Prepare("INSERT INTO `go_short_url_rule` (`suffix`, `target`, `user_id`, `expires`) VALUES (?, ?, ?, ?)")
 	if err != nil {
 		log.Fatalln("[(rule Rule) Insert-1]", err)
@@ -35,16 +36,16 @@ func (rule Rule) Insert(db *sql.DB) (r *Rule, err error) {
 	}
 	insertId, err := util.ExecErrorHandler(stmt.Exec(rule.Suffix, rule.Target, rule.UserId, expiresString))
 	if err != nil {
-		return nil, err
+		return err
 	} else {
 		rule.Id = insertId
 		rule.CreateTime = time.Now()
 		rule.UpdateTime = time.Now()
-		return &rule, nil
+		return nil
 	}
 }
 
-func (rule Rule) Update(db *sql.DB) error {
+func (rule *Rule) Update(db *sql.DB) error {
 	stmt, err := db.Prepare("UPDATE `go_short_url_rule` SET `target` = ?, `request` = ?, `user_id` = ?, `expires` = ? WHERE `id` = ?")
 	if err != nil {
 		return err
@@ -61,8 +62,10 @@ func (rule Rule) Update(db *sql.DB) error {
 	return nil
 }
 
-func SelectTargetBySuffix(db *sql.DB, suffix string) (*Rule, error) {
-	stmt, err := db.Prepare("SELECT `id`, `suffix`, `target`, `request`, `user_id`, `expires`, `create_time`, `update_time` FROM `go_short_url_rule` WHERE `suffix` = ?")
+const SELECT_PREFIX = "SELECT `id`, `suffix`, `target`, `request`, `user_id`, `expires`, `create_time`, `update_time` FROM `go_short_url_rule`"
+
+func SelectRuleBySuffix(db *sql.DB, suffix string) (*Rule, error) {
+	stmt, err := db.Prepare(SELECT_PREFIX + " WHERE `suffix` = ?")
 	if err != nil {
 		return nil, err
 	}
@@ -87,14 +90,45 @@ func SelectTargetBySuffix(db *sql.DB, suffix string) (*Rule, error) {
 	return &rule, nil
 }
 
-func DeleteById(db *sql.DB, id int64) error {
-	stmt, err := db.Prepare("DELETE FROM `go_short_url_rule` WHERE `id` = ?")
+func DeleteById(db *sql.DB, userId int64, id int64) error {
+	stmt, err := db.Prepare("DELETE FROM `go_short_url_rule` WHERE `id` = ? AND `user_id` = ?")
 	if err != nil {
 		log.Fatalln("[DeleteById]", err)
 	}
-	_, err = util.ExecErrorHandler(stmt.Exec(id))
+	_, err = util.ExecErrorHandler(stmt.Exec(id, userId))
 	if err != nil {
 		return err
 	}
 	return nil
+}
+
+func SearchRules(db *sql.DB, userId int64, keyword string, page int, pageSize int) ([]Rule, error) {
+	stmt, err := db.Prepare(SELECT_PREFIX + " WHERE CONCAT(`suffix`, `target`) LIKE ? AND `user_id` = ? LIMIT ? OFFSET ?")
+	if err != nil {
+		log.Fatalln("[SearchRules-1]", err)
+	}
+	rows, err := stmt.Query("%"+strings.Join(strings.Split(keyword, " "), "%")+"%", userId, pageSize, page*pageSize)
+	if err != nil {
+		return nil, err
+	}
+	var rules []Rule
+	defer rows.Close()
+	for rows.Next() {
+		rule := Rule{}
+		createTime := ""
+		updateTime := ""
+		expires := sql.NullString{}
+		err = rows.Scan(&rule.Id, &rule.Suffix, &rule.Target, &rule.Request, &rule.UserId, &expires, &createTime, &updateTime)
+		if err != nil {
+			log.Fatalln(err)
+		}
+		rule.CreateTime = util.ParseTimeFromDB(createTime)
+		rule.UpdateTime = util.ParseTimeFromDB(updateTime)
+		if expires.Valid {
+			datetime := util.ParseTimeFromDB(expires.String)
+			rule.Expires = &datetime
+		}
+		rules = append(rules, rule)
+	}
+	return rules, nil
 }
